requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
sqlalchemy
langchain
langchain-community
faiss-cpu
pymupdf
pdfplumber
pandas
openpyxl
sentence-transformers
httpx
requests
jinja2
